fixed problem 2 in c++

+++ lib/c++/factor.h
@@ -1,10 +1,8 @@
 #ifndef factor_h__
 #define factor_h__
 
-#include <stdbool.h>
-
 bool even(int n);
-bool even_l(int n);
+bool even_l(long n);
 
 bool odd(int n);
 bool odd_l(long n);
